fixed parameter for model, updated view methods

+++ src/cellsociety/Model.java
@@ -7,7 +7,7 @@ public class Model {
 
   }
 
-  public Cell getCell(int x, int y){
+  public Cell getCell(double x, double y){
 
   }
